Flask-OCR-Skincare-Analysis/requirements.txt
Flask==2.3.2
tensorflow==2.12.0
easyocr==1.6.2
pandas==1.5.3
Pillow==9.4.0
numpy==1.24.2
